Remove image casting into double from loadImage method
- Modify common function calling loadImage
- Add method to generate channel file names with single files

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@8513 0bcb941b-74d1-0310-bb26-e6636707f451

+++ detectionAlgorithms/subResFeatures2D/detectSubResFeatures2D_StandAlone.m
@@ -342,7 +342,7 @@ for iWindow = 1 : numIntegWindow
                 end
             end
         else
-            imageRaw=channel.loadImage(imageIndx(iImage:iImage+2*integWindow(iWindow)));
+            imageRaw=double(channel.loadImage(imageIndx(iImage:iImage+2*integWindow(iWindow))));
         end
         %replace zeros with NaNs
         %zeros result from cropping that leads to curved boundaries
